Create substring method
The substring method gets a string and check if there are substrings
of the received string in an array. The return is a Hash, containing
How many times a substring appeared in the given string

First I used split method to turn the given string into an array of
strings. This is useful for multiple word strings.

Then I iterated through the array, using reduce method to create a hash.
Inside the array iteration we iterate through the newly created array of
strings, checking if any match with the include? method. If match,
the Hash for that specific substring is incremented by one.

The end result is the Hash with the substrings that appear in the given
string, and how many times each substring appears.